improve the io to allow png and jpg and to allow the export of files as h5 files
