+++ plantseg/io/__init__.py
@@ -1,3 +1,4 @@
-from plantseg.io.io import allowed_data_format, H5_EXTENSIONS, TIFF_EXTENSIONS
-from plantseg.io.io import read_tiff_voxel_size, read_h5_voxel_size
-from plantseg.io.io import smart_load, create_tiff, create_h5, load_shape
+from plantseg.io.io import smart_load, load_shape, allowed_data_format
+from plantseg.io.tiff import load_tiff, read_tiff_voxel_size, create_tiff, TIFF_EXTENSIONS
+from plantseg.io.h5 import load_h5, read_h5_voxel_size, create_h5, H5_EXTENSIONS
+from plantseg.io.pil import load_pill, PIL_EXTENSIONS
